elbowEigen {SPAC2}	R Documentation

Alternative methods based on the detection of elbow in sample eigenvalues.

Description

The function returns the dimension that is selected based on various methods that attempt to detect an “elbow” in sample eigenvalues.

Usage


elbowEigen(lambda, methods = c("adjD", "cumD", "varD", "cumlog", "logsigma2"))

Arguments

lambda
a numeric vector of positive sample eigenvalues of length n

methods
a character vector of possible methods to detect an elbow in sample eigenvalues.

Value

an integer $K$ between 1 and $n$.

Examples


## Not run: 
library(MASS)
X <- mvrnorm(1000, mu = rep(0,10), Sigma = diag(1,10))
eigen_values <- eigen(as.matrix(Matrix::nearPD(stats::cov(scale(X)))$mat))$val
elbowEigen(lambda = eigen_values)
elbowEigen(lambda = eigen_values)

## End(Not run)



[Package SPAC2 version 0.9.9 Index]