loglk {SPAC2}	R Documentation

Profile log-likelihood of the PPCA model with sample eigenvalues.

Description

The function returns the profile log-likelihood of the PPCA model for each possible choice of K (1, 2, …, n-1) at their respective MLEs. The maximum choice was set at n-1 because when K=n, the profile log-likelihood is equal to that at K=n-1.

Usage


loglk(lam, n, tau = 1e-05)

Arguments

lam
a numerical vector of positive sample eigenvalues

n
the full dimension

tau
a tolerance threshold for the smallest eigenvalue, the default value is 0.001.

Value

profile log-likelihood of length n-1.

[Package SPAC2 version 0.9.9 Index]