pPPCA_INFO {SPAC2}	R Documentation

Penalized profile log-likelihood of a PPPCA model

Description

The function returns either the penalized profile log-likelihood or the value that maximizes the penalized profile log-likelihood for a given tuning parameter value.

Usage


pPPCA_INFO(lam, Tvotes = 100, M, AIC = TRUE, tau = 0.001, beta = NULL)

Arguments

lam
a numerical vector of positive sample eigenvalues

Tvotes
the number of possible tuning parameter values to be searched

M
the number of observations used to calculate sample covariance.

AIC
a logical indicator to use AIC as the information criterion, if FALSE, then BIC is used; the default option is AIC.

tau
a tolerance threshold for the smallest eigenvalue, the default value is 0.001.

beta
a numeric between 0 and 1 indicating the weight towards penalty function 1 or 2.

Value

an integer K that maximizes the penalized profile log-likelihood for the given delta value.

[Package SPAC2 version 0.9.9 Index]